requirements.txt
# Web Framework & Server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]==2.6.1

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
cryptography==42.0.2

# Email
email-validator==2.1.0.post1
aiosmtplib==2.0.2

# Machine Learning & NLP
transformers==4.37.2
torch==2.2.0
torchaudio==2.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
nltk==3.8.1
spacy
textblob==0.17.1
deepface
tf-keras

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
SpeechRecognition==3.10.1
pyaudio==0.2.14

# Video Processing
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
mediapipe==0.10.8
moviepy==1.0.3

# Testing & Development
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.39.1
structlog==24.1.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2024.1
uuid==1.30

# Post-installation requirements
# Run these commands after installation:
# python -m spacy download en_core_web_sm
# python -m nltk.downloader punkt
# python -m nltk.downloader averaged_perceptron_tagger
# python -m nltk.downloader wordnet 
